+++ rotkehlchen/db/upgrade_manager.py
@@ -276,7 +276,7 @@ class DBUpgradeManager():
             try:
                 kwargs = upgrade.kwargs if upgrade.kwargs is not None else {}
                 upgrade.function(db=self.db, **kwargs)
-            except BaseException as e:
+            except BaseException as e:  # lgtm[py/catch-base-exception]
                 # Problem .. restore DB backup and bail out
                 error_message = (
                     f'Failed at database upgrade from version {upgrade.from_version} to '