+++ rotkehlchen/rotkehlchen.py
@@ -15,7 +15,6 @@ from rotkehlchen.accounting.accountant import Accountant
 from rotkehlchen.accounting.structures import Balance, BalanceType
 from rotkehlchen.api.websockets.notifier import RotkiNotifier
 from rotkehlchen.api.websockets.typedefs import WSMessageType
-from rotkehlchen.app_upgrades.manager import RotkiUpgradeManager
 from rotkehlchen.assets.asset import Asset
 from rotkehlchen.balances.manual import (
     account_for_manually_tracked_asset_balances,
@@ -38,6 +37,7 @@ from rotkehlchen.config import default_data_directory
 from rotkehlchen.constants.misc import ZERO
 from rotkehlchen.data.importer import DataImporter
 from rotkehlchen.data_handler import DataHandler
+from rotkehlchen.data_migrations.manager import DataMigrationManager
 from rotkehlchen.db.settings import DBSettings, ModifiableDBSettings
 from rotkehlchen.errors import (
     EthSyncError,
@@ -186,7 +186,7 @@ class Rotkehlchen():
         self.premium_sync_manager = PremiumSyncManager(data=self.data, password=password)
         # set the DB in the external services instances that need it
         self.cryptocompare.set_database(self.data.db)
-        RotkiUpgradeManager(self).run_upgrades()
+        DataMigrationManager(self).maybe_migrate_data()
 
         # Anything that was set above here has to be cleaned in case of failure in the next step
         # by reset_after_failed_account_creation_or_login()