+++ rotkehlchen/tests/fixtures/db.py
@@ -114,6 +114,7 @@ def _init_database(
         include_cryptocompare_key: bool,
         tags: List[Dict[str, Any]],
         manually_tracked_balances: List[ManuallyTrackedBalance],
+        data_migration_version: int,
 ) -> DBHandler:
     db = DBHandler(
         user_data_dir=data_dir,
@@ -122,7 +123,7 @@ def _init_database(
         initial_settings=None,
     )
     # Make sure that the fixture provided data are included in the DB
-    add_settings_to_test_db(db, db_settings, ignored_assets)
+    add_settings_to_test_db(db, db_settings, ignored_assets, data_migration_version)
     add_blockchain_accounts_to_db(db, blockchain_accounts)
     maybe_include_etherscan_key(db, include_etherscan_key)
     maybe_include_cryptocompare_key(db, include_cryptocompare_key)
@@ -145,6 +146,7 @@ def database(
         include_cryptocompare_key,
         tags,
         manually_tracked_balances,
+        data_migration_version,
 ) -> Optional[DBHandler]:
     if not start_with_logged_in_user:
         return None
@@ -160,6 +162,7 @@ def database(
         include_cryptocompare_key=include_cryptocompare_key,
         tags=tags,
         manually_tracked_balances=manually_tracked_balances,
+        data_migration_version=data_migration_version,
     )
 
 
@@ -175,6 +178,7 @@ def session_database(
         session_include_cryptocompare_key,
         session_tags,
         session_manually_tracked_balances,
+        data_migration_version,
 ) -> Optional[DBHandler]:
     if not session_start_with_logged_in_user:
         return None
@@ -192,6 +196,7 @@ def session_database(
         include_cryptocompare_key=session_include_cryptocompare_key,
         tags=session_tags,
         manually_tracked_balances=session_manually_tracked_balances,
+        data_migration_version=data_migration_version,
     )
 
 
@@ -202,4 +207,9 @@ def fixture_db_settings() -> Optional[Dict[str, Any]]:
 
 @pytest.fixture(scope='session', name='session_db_settings')
 def fixture_session_db_settings() -> Optional[Dict[str, Any]]:
+    return None
+
+
+@pytest.fixture(name='use_custom_database')
+def fixture_use_custom_database() -> Optional[str]:
     return None